Update NHLService to return NHL stats as they are received
This update was made because NHL updated their team statistics endpoint
to use the same service as the skaters or players endpoint.  This is
a much simpler endpoint.  However we do lose the team conference and
division.

This is a major release because the NHL.teamStats method now returns a
Map<String, Object> rather than Collection<Map<String, Object>>.

Update version to 3.0.0